backend/requirements.txt
fastapi
uvicorn[standard]
redis
httpx
python-dotenv
langgraph
langchain-core
pydantic
fakeredis
pymongo
razorpay